temporarily run tests in forked jvm to avoid classpath issues
scala-parser-combinators 1.1.2 and scala 2.11 do not get along well in tests (because scala 2.11 has
a different version of scala-parser-combinators on its classpath). forking the process when running
the tests means that the tests can have their own, separate classpath which will not have version
conflicts from this lib. Scalas >= 2.12 do not have scala-parser-combinators on their classpath,
so this problem will go away when we drop Scala 2.11 (hopefully soon)