remove play-guice from atom-manager-play-lib
We're only using it for convenience in tests, and it doesn't form part of the API so no
need to be installed in dependents, and play-test already brings in play-guice in the test scope
so no need to even explicitly depend on it. This removes the last symbol from the play plugin, so
we can remove that too, which means the Play version is only defined in one place so cannot go out
of sync again.